Move species components to folder
